+++ benchmark/seq_hdt/dynamic_graph/include/dynamic_graph/graph.hpp
@@ -10,7 +10,7 @@
 #include <utility>
 
 /** Represents a vertex in a graph. */
-typedef int64_t Vertex;
+typedef uint32_t Vertex;
 
 /** Represents an edge in a directed graph. */
 typedef std::pair<Vertex, Vertex> DirectedEdge;
